remarks fixes
